Fix in Pair.h

+++ src/Object.h
@@ -188,7 +188,7 @@ public:
     static Object makeBignum(signed long int n);
     static Object makeBignum(Bignum* b);
     static Object makeFlonum(double value);
-    static Object makeRaw(int n);
+    static Object makeRaw(intptr_t n);
     static Object makeRaw(void* n);
     static Object makeInstruction(int n);
     static Object makeCompilerInstruction(int n);
